Add ISQLiteContext to manage connection and transaction usage in SQLite event store
There are two modes of operation, default and manual. In DefaultSQLiteContext,
a connection is opened on-demand and transactions are created and committed
or rolled back immediately. In ManualSQLiteContext, connection and transaction
management are entirely up to the client. This allows the usage of ATTACH
commands to perform atomic commits across multiple SQLite databases in a
single transaction.