Created a Castle Windsor based CommandService
Allows command executors and interceptors to be resolved directly from the
Windsor container instead of depending on manual registration.

Note: This require a minor change to the CommandContext class.  The constructor
had to be extended with optional parameters to allow extension since all the
setters have been marked internal.  This maintains the immutability of the type
while leaving it more open for extension.