Updated Version of MyNotes
MyNotes 2012 is an updated/re-worked version of the MyNotes sample that
goes in a slightly different direction.  Rather than have two
databases--one for the event store and one for the read model--MyNotes
2012 focuses on having a single database, but with the possibility for
separate schemas.  MyNotes 2012 also updates the technology stack
slightly--eliminating edmx files in the read model in favor of POCOs,
using MVC4 for the web UI, and setting the root namespace for all the
assembly projects to "MyNotes".