NSB EventMessage Changes
Ok so with NSB 4 you can no longer have generic type messages. This
meant the original implementation of EventMessage<T> was not valid.

For our companies implementation we had a single NSB Message Handler for
EventMessage and then put the Event payload onto an InProcessEventBus
where the actual NCQRS IEventHandlers took care of the event handling.
This simplified the NSB configuration as we only have to register one
type (EventMessage) as a NSB event.