Batching Commands
Ok this change has a bit of code smell in our project, however it is
absolutely giving the desired results even tho we believe it's getting
there incorrectly.

Our domain has lots of small aggregates and some of our user workflows
involve the creation of many of these aggregates at once in various
combinations.

We ended up having a 'batch' of create commands that we wanted to submit
as a single transaction. I.e. if any of the commands failed for whatever
reason the entire batch of commands failed.

We do this by having creating a CommandMessage with a payload of
ICommands. When our NSB-Application Service processes this single
message it just loops through the commands for execution.

In our environment this is working and giving the desired results...
however we THINK this is happening not because of some elegant UoW
transaction handling in NCQRS (because that only exists on a per command
basis) but instead on an NSB - SQL SERVER - DTC magic.

For now we're happy with this results but will investigate a 'better
way' to achieve this functionality within the framework.